requirements.txt
python-dotenv
telethon=1.34.0